Refactor database imports to use single import from database module
