+++ frontend/components/experiment_details.py
@@ -4,16 +4,7 @@ import datetime
 import os
 import json
 from sqlalchemy import orm
-from database.database import SessionLocal
-from database.models import (
-    SampleInfo,
-    ExternalAnalysis,
-    ExperimentNotes,
-    Experiment,
-    ExperimentStatus,
-    ExperimentalConditions,
-    PXRFReading,
-)
+from database import SessionLocal, SampleInfo, ExternalAnalysis, ExperimentNotes, Experiment, ExperimentStatus, ExperimentalConditions, PXRFReading
 from frontend.config.variable_config import (
     FIELD_CONFIG,
     EXPERIMENT_TYPES,
