+++ tests/test_load_info.py
@@ -9,7 +9,7 @@ from datetime import datetime
 project_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
 sys.path.insert(0, project_root)
 
-from database.models import Base, SampleInfo, ExternalAnalysis, PXRFReading
+from database import Base, SampleInfo, ExternalAnalysis, PXRFReading
 from frontend.config.variable_config import ELEMENT_COLUMNS # Need this for checks
 # Import the function to test
 from frontend.components.load_info import get_external_analyses
